Prepare interfaces and fix issues needed by CDE Editor
CDE Editor will need the extended versions of IUserContentAccess and
IContentAccessFactory.
