+++ core/src/main/java/pt/webdetails/cpf/api/IContentAccessFactoryExtended.java
@@ -1,0 +1,28 @@
+/*!
+ * Copyright 2002 - 2018 Webdetails, a Hitachi Vantara company.  All rights reserved.
+ *
+ * This software was developed by Webdetails and is provided under the terms
+ * of the Mozilla Public License, Version 2.0, or any later version. You may not use
+ * this file except in compliance with the license. If you need a copy of the license,
+ * please go to  http://mozilla.org/MPL/2.0/. The Initial Developer is Webdetails.
+ *
+ * Software distributed under the Mozilla Public License is distributed on an "AS IS"
+ * basis, WITHOUT WARRANTY OF ANY KIND, either express or  implied. Please refer to
+ * the license for the specific language governing your rights and limitations.
+ */
+package pt.webdetails.cpf.api;
+
+import pt.webdetails.cpf.repository.api.IContentAccessFactory;
+
+/**
+ * Minimal repository access provider for basic plugin needs 
+ */
+public interface IContentAccessFactoryExtended extends IContentAccessFactory {
+
+  /**
+   * @param basePath (optional) all subsequent paths will be relative to this
+   * @return {@link IUserContentAccess} for user repository access
+   */
+  IUserContentAccessExtended getUserContentAccess( String basePath );
+
+}
