+++ core/src/main/java/pt/webdetails/cpf/packager/dependencies/FileDependency.java
@@ -1,6 +1,15 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+/*!
+ * Copyright 2013-2018 Webdetails, a Hitachi Vantara company. All rights reserved.
+ *
+ * This software was developed by Webdetails and is provided under the terms
+ * of the Mozilla Public License, Version 2.0, or any later version. You may not use
+ * this file except in compliance with the license. If you need a copy of the license,
+ * please go to http://mozilla.org/MPL/2.0/. The Initial Developer is Webdetails.
+ *
+ * Software distributed under the Mozilla Public License is distributed on an "AS IS"
+ * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. Please refer to
+ * the license for the specific language governing your rights and limitations.
+ */
 
 package pt.webdetails.cpf.packager.dependencies;
 
@@ -63,7 +72,12 @@ public class FileDependency extends Dependency {
   }
 
   public InputStream getFileInputStream() throws IOException {
-    return origin.getReader( getContentFactory() ).getFileInputStream( filePath );
+    if ( getContentFactory() != null ) {
+      return origin.getReader( getContentFactory() ).getFileInputStream( filePath );
+    } else {
+      logger.fatal( String.format( "Couldn't getFileInputStream() for filePath = '%s'. Unable to get ContentFactory.", filePath ) );
+      return null;
+    }
   }
 
   /**
@@ -86,7 +100,11 @@ public class FileDependency extends Dependency {
   }
 
   protected IContentAccessFactory getContentFactory() {
-    return PluginEnvironment.env().getContentAccessFactory();
+    if ( PluginEnvironment.env() != null ) {
+      return PluginEnvironment.env().getContentAccessFactory();
+    } else {
+      return null;
+    }
   }
 
   public String toString() {
