--- pentaho/src/main/java/pt/webdetails/cpf/api/IFileContent.java
@@ -1,22 +1,0 @@
-package pt.webdetails.cpf.api;
-
-import pt.webdetails.cpf.repository.api.IBasicFile;
-
-public interface IFileContent extends IBasicFile {
-
-  /**
-   * @return the title of the file
-   */
-  String getTitle();
-
-  /**
-   * @return the description of the file
-   */
-  String getDescription();
-
-  /**
-   * @return flag hidden
-   */
-  boolean isHidden();
-
-}
